requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.1
django-cors-headers==4.3.1
django-crontab==0.7.1
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-filter==24.2
django-utils-six==2.0
djangorestframework==3.15.1
djangorestframework-camel-case==1.4.2
drf-yasg==1.21.7
Faker==24.4.0
geographiclib==2.0
geopy==2.4.1
gunicorn==21.2.0
inflection==0.5.1
kombu==5.3.6
packaging==24.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
